add async await to todo api

+++ server/server.js
@@ -56,23 +56,21 @@ app.get('/todos/:id', authenticate, (req, res) => {
   }).catch((e) => res.status(400).send());
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-  var id = req.params.id;
-  if(!ObjectID.isValid(id)){
-    return res.status(404).send();
+app.delete('/todos/:id', authenticate, async (req, res) => {
+  try{
+    var id = req.params.id;
+    if(!ObjectID.isValid(id)){
+      return res.status(404).send();
+    }
+  
+    var todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id});
+    if(!todo){
+      return res.status(404).send();
+    }
+    res.send({todo});
+  } catch (e){
+    res.status(400).send();
   }
-
-  Todo.findOneAndRemove({
-      _id: id, 
-      _creator: req.user._id
-    })
-    .then((todo) => {
-      if(!todo){
-        return res.status(404).send();
-      }
-
-      res.send({todo});
-    }).catch((e) => res.status(400).send());
 });
 
 app.patch('/todos/:id', authenticate, (req, res) => {
@@ -103,15 +101,17 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 
 });
 
-app.post('/users', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-  var user = new User(body);
-  user.save().then(() => {
-      return user.generateAuthToken();
-    }).then((token) => {
-      res.header('x-auth', token).send({user});
-    }).catch((e) => res.status(400).send(e));
-  });
+app.post('/users', async (req, res) => {
+  try{
+    var body = _.pick(req.body, ['email', 'password']);
+    var user = new User(body);
+    await user.save();
+    var token = await user.generateAuthToken();
+    res.header('x-auth', token).send({user});
+  } catch(e) {
+    res.status(400).send(e);
+  }
+});
 
 app.get('/users/me', authenticate, (req, res) => {
   res.send({
@@ -119,22 +119,25 @@ app.get('/users/me', authenticate, (req, res) => {
   });
 });
 
-app.post('/users/login', (req, res) => {
-  var {email, password} = req.body;
-  User.findByCredentials(email, password).then((user) => {
-    return user.generateAuthToken().then((token) => {
-      res.header('x-auth', token).send({user});
-    });
-  }).catch((e) => res.status(400).send(e));
+app.post('/users/login', async (req, res) => {
+  try{
+    var {email, password} = req.body;
+    var user = await User.findByCredentials(email, password);
+    var token = await user.generateAuthToken();
+    res.header('x-auth', token).send({user});
+  } catch(e){
+    res.status(400).send(e);
+  }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
   var {user, token} = req;
-  user.removeToken(token).then(() => {
+  try {
+    await user.removeToken(token);
     res.status(200).send();
-  }, (e) => {
+  } catch (e) {
     res.status(400).send();
-  })
+  }
 });
   
 module.exports = {
